+++ packages/studio-base/src/components/AppBar/index.tsx
@@ -23,7 +23,7 @@ import {
   CustomWindowControls,
   CustomWindowControlsProps,
 } from "@foxglove/studio-base/components/AppBar/CustomWindowControls";
-import AppMenu from "@foxglove/studio-base/components/AppMenu";
+import { AppMenu } from "@foxglove/studio-base/components/AppMenu";
 import { FoxgloveLogo } from "@foxglove/studio-base/components/FoxgloveLogo";
 import { MemoryUseIndicator } from "@foxglove/studio-base/components/MemoryUseIndicator";
 import Stack from "@foxglove/studio-base/components/Stack";
