+++ packages/studio-base/src/components/NestedMenuItem.tsx
@@ -1,0 +1,142 @@
+// This Source Code Form is subject to the terms of the Mozilla Public
+// License, v2.0. If a copy of the MPL was not distributed with this
+// file, You can obtain one at http://mozilla.org/MPL/2.0/
+
+import { ChevronRight12Regular } from "@fluentui/react-icons";
+import { Divider, Menu, MenuItem } from "@mui/material";
+import {
+  Dispatch,
+  MouseEvent,
+  PropsWithChildren,
+  ReactNode,
+  SetStateAction,
+  useState,
+} from "react";
+import { makeStyles } from "tss-react/mui";
+
+export type NestedMenuItem =
+  | {
+      type: "item";
+      label: ReactNode;
+      key: string;
+      disabled?: boolean;
+      shortcut?: string;
+      onClick?: () => void;
+    }
+  | { type: "divider" };
+
+const useStyles = makeStyles<void, "icon">()((theme, _params, classes) => ({
+  nestedMenuPaper: {
+    marginTop: theme.spacing(-1),
+  },
+  menuItem: {
+    justifyContent: "space-between",
+    cursor: "pointer",
+    gap: theme.spacing(2),
+
+    "&.Mui-selected, &.Mui-selected:hover": {
+      backgroundColor: theme.palette.action.hover,
+    },
+    [`:not(:hover, :focus) .${classes.icon}`]: {
+      opacity: 0.6,
+    },
+    kbd: {
+      font: "inherit",
+      color: theme.palette.text.disabled,
+    },
+  },
+  menuList: {
+    minWidth: 180,
+    maxWidth: 220,
+  },
+  icon: {},
+  endIcon: {
+    marginRight: theme.spacing(-0.5),
+  },
+}));
+
+export function NestedMenuItem(
+  props: PropsWithChildren<{
+    items: NestedMenuItem[];
+    subMenu?: HTMLElement;
+    setSubMenu: Dispatch<SetStateAction<HTMLElement | undefined>>;
+    subMenuOpen: boolean;
+    id?: string;
+  }>,
+): JSX.Element {
+  const { classes, cx } = useStyles();
+  const { children, items, subMenu, subMenuOpen, setSubMenu, id } = props;
+  const [anchorEl, setAnchorEl] = useState<undefined | HTMLElement>(undefined);
+  const open = Boolean(anchorEl);
+
+  const handleClick = (event: MouseEvent<HTMLElement>) => {
+    if (anchorEl !== event.currentTarget) {
+      setSubMenu(event.currentTarget);
+      setAnchorEl(event.currentTarget);
+    }
+  };
+
+  const handleMouseEnter = (event: MouseEvent<HTMLElement>) => {
+    setSubMenu(event.currentTarget);
+    setAnchorEl(event.currentTarget);
+  };
+
+  const handleMouseLeave = (event: MouseEvent<HTMLElement>) => {
+    if (!subMenuOpen && subMenu !== event.currentTarget) {
+      setAnchorEl(undefined);
+    }
+    setSubMenu(undefined);
+  };
+
+  return (
+    <>
+      <MenuItem
+        id={id}
+        selected={open}
+        className={classes.menuItem}
+        onClick={handleClick}
+        onMouseEnter={handleMouseEnter}
+        onMouseLeave={handleMouseLeave}
+        data-testid={id}
+      >
+        {children}
+        <ChevronRight12Regular className={cx(classes.icon, classes.endIcon)} />
+      </MenuItem>
+      <Menu
+        open={open}
+        disablePortal
+        anchorEl={anchorEl}
+        onClose={() => setAnchorEl(undefined)}
+        onMouseLeave={() => {
+          setAnchorEl(undefined);
+          setSubMenu(undefined);
+        }}
+        anchorOrigin={{ vertical: "top", horizontal: "right" }}
+        MenuListProps={{ dense: true, className: classes.menuList }}
+        autoFocus={false}
+        disableAutoFocus
+        disableEnforceFocus
+        hideBackdrop
+        PaperProps={{
+          className: classes.nestedMenuPaper,
+        }}
+      >
+        {items.map((item, idx) =>
+          item.type !== "divider" ? (
+            <MenuItem
+              className={classes.menuItem}
+              key={item.key}
+              onClick={item.onClick}
+              disabled={item.disabled}
+            >
+              {item.label}
+              {item.shortcut && <kbd>{item.shortcut}</kbd>}
+            </MenuItem>
+          ) : (
+            <Divider key={`${idx}-divider`} variant="middle" />
+          ),
+        )}
+      </Menu>
+    </>
+  );
+}
